Essentially the minimum viable product
This module provides a direct way to hook the schedule function
so that (at the moment) a callback can b e invoked just after a
context switch. The module also contains a mechanism to follow
thread creation and death so that the functionality can be
extended also to the children of the registred processes.