Improve overall support.
This commit enables the process_exit tracker so that it is possible to monitor
the number of current tracked processes. Furthermore, it is possible to define a
callback to be executed when the number of processes reached 0.

The system_hook module checks, at the context switch, the process' tgid to run
the ctx_callback. This makes it easier to track the process's children.
The thread-safety has been improved by a reference counter and grace-period-like mechanism that assures that no core is executing the ctx_callback while the
provider module is being unmounted.

Fix minor bugs.