Adapt to Linux Kenrel 5.11
The primary target of this commit is to make the kprobe symbols
compatible with different Linux Kernel versions, up to 5.11.
The entire module has been refactorized to make functions' names
fit their scope.
There is preliminary support to track threads via TGID, but it is not
working correctly, so it used the PID by default.
The tracked threads list has been converted into a hashtable to improve
performance when querying the module.