Define per-PID and per-TGID tracking.
It is possible to set the way each process should be tracked.
SHOOK will track each process independently if the PID is selected,
while by choosing the TGID, the entire group will be tracked instead.
In the latter case, the TGID entry is removed only when the registered
process counter reaches the 0 value.